+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [UNRELEASED]
+
+### Fixed
+
+- SQL expression in supplier survey total calculation
+
 ## [2.11.2] - 2025-08-07
 
 - Update exemple.odt file to include ecotax tags