+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 
 - SQL expression in supplier survey total calculation
+- Add `users_id_tech` field field for tech assignment compatibility
 
 ## [2.12.0] - 2025-10-01